Improve coverage in test for form examples
In the test for form examples, the tests are not testing the
route that displays the form. The tests are only testing the
result of the form submission. As a result, not all code in the
tests is being executed. Ensure that 100% of the test code is
executed by testing the route that displays the form.