Add test for empty PATH_INFO in environ
There is a check in ice.Request that tests if the value for
PATH_INFO key in environ dictionary is an empty string. If it is
found to be an empty string, the value is set to '/'. This check
has not been tested in the tests for this class. As a result,
the ice module does not have 100% test coverage. Fix this by
adding a test for this check.