Rudimentary error message when no error handler
If an ice application is created directly using the ice.Ice
class, i.e. app = ice.Ice(), and no fallback error handler is
defined for this app, then an HTTP error leads to a blank web
page because the response body is an empty string. With this
change, the response body is no longer an empty string for such
a case. A rudimentary error message with the HTTP status code
and reason phrase is displayed instead.