Let status code be returned from route's callable
With this change, when an integer representing an HTTP response
status code is returned from a route's callable, an HTTP
response with this status code is sent to the client. While
doing so, if the response body is found to be empty, a default
response body is created with error handler.