Add wildcard of path type
Add support for wildcard of path type that matches one or more
characters that may contain slash, i.e. /. This allows a portion
of a path to be matched by the wildcard. Wildcard of str type
cannot do this because such a wildcard does not match a string
that contains slash.