Add static() method to return static files
Add static() method to Ice class that may be used to return
content of static files.

Explain in the documentation that this is a very rudimentary
means of serving static files which could be useful in a
development environment where one would want to test pages with
static content such as style sheets, images, etc. served by an
ice application without using a web server.