Make the cookie test work with Python 3.4.3 and later versions
The test case for cookies fails with Python 3.4.3 and later versions
because of a difference in how the cookie attributes are normalized in
these versions and earlier versions.

In earlier versions, these cookie attributes are normalized to
"httponly" and "secure" whereas in the later versions, these cookie
attributes are normalized to "HttpOnly" and "Secure". See the following
URLs for more details.

  - https://docs.python.org/3.4/whatsnew/changelog.html
  - http://bugs.python.org/issue23250

Therefore, modify the test case for cookies such that it matches these
cookie attributes in a case-insensitive manner.